requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
redis>=5.0.0
numpy>=1.24.0
pandas>=2.0.0 